Standardize `getLevelBySeverity` functions
